I've reworked the entire intepreter by wrapping Symex.Result into a double-state monad.
At first I was skeptical, especially when rewriting the first few functions.
It seemed to clutter a lot of types, and I was afraid of perfs.
However, it turned out magnificent - in my opinion. For instance, interpretation of loops:

Before:
```ocaml
| AilSwhile (cond, stmt, _loopid) ->
    let rec loop store state =
      let** cond_v, store, state = eval_expr ~prog store state cond in
      if%sat cast_to_bool cond_v then
        let** res, store, state = exec_stmt ~prog store state stmt in
        match res with
        | Some _ -> Result.ok (res, store, state)
        | None -> loop store state
      else Result.ok (None, store, state)
    in
    loop store state
```

After:
```ocaml
| AilSwhile (cond, stmt, _loopid) ->
    let rec loop () =
      let* cond_v = eval_expr ~prog cond in
      if%sat cast_to_bool cond_v then
        let* res = exec_stmt ~prog stmt in
        match res with Some _ -> InterpM.ok res | None -> loop ()
      else InterpM.ok None
    in
    loop ()
```

This is much more readable, and as soon as I make `prog` accessible through an effect (next PR, already done), the apparent signature of `exec_stmt` will literally be:
```ocaml
val eval_expr: Stmt.t -> Value.t option
```
which is really cool.

Perf-wise it doesn't change anything:
```
Before Interp Monad:
Time (mean ± σ):      5.153 s ±  0.038 s    [User: 1.601 s, System: 0.794 s]
Range (min … max):    5.091 s …  5.218 s    10 runs

After Interp Monad:
  Time (mean ± σ):      5.147 s ±  0.035 s    [User: 1.609 s, System: 0.798 s]
  Range (min … max):    5.086 s …  5.199 s    10 runs
```